requirements.txt
delorean
colander
beautifulsoup4==4.3.2
Django==1.7.8
django-appconf==1.0.1
django-compressor==1.5
django-debug-toolbar==1.3.0
django-libsass==0.3
django-modelcluster==0.6.2
django-taggit==0.12.3
django-treebeard==3.0
html5lib==0.999
libsass==0.8.0
Pillow==2.8.1
pytz==2015.2
requests==2.7.0
six==1.9.0
sqlparse==0.1.15
Unidecode==0.4.17
wagtail==1.0b1
wagtailnews==0.3.0
Willow==0.1
